Reorganize cgroup modules
