+++ src/cgroups/mod.rs
@@ -1,13 +1,3 @@
-mod controller;
-mod controller_type;
-mod devices;
-mod hugetlb;
-mod blkio;
-mod manager;
-mod memory;
-mod network_classifier;
-mod network_priority;
-mod pids;
-pub use controller::Controller;
-pub use controller_type::ControllerType;
-pub use manager::Manager;
+pub mod v1;
+pub mod v2;
+
