+++ src/cgroups/v1/devices.rs
@@ -8,7 +8,7 @@ use anyhow::Result;
 use nix::unistd::Pid;
 
 use crate::{
-    cgroups::Controller,
+    cgroups::v1::Controller,
     rootfs::default_devices,
 };
 use oci_spec::{LinuxDeviceCgroup, LinuxDeviceType, LinuxResources};
