+++ src/cgroups/v2/cpu.rs
@@ -1,0 +1,3 @@
+pub struct Cpu {
+    
+}
