+++ src/main.rs
@@ -14,7 +14,7 @@ use youki::create;
 use youki::signal;
 use youki::start;
 use youki::utils;
-use youki::{cgroups::Manager, command::linux::LinuxCommand};
+use youki::{cgroups::v1::Manager, command::linux::LinuxCommand};
 
 /// High-level commandline option definition
 /// This takes global options as well as individual commands as specified in [OCI runtime-spec](https://github.com/opencontainers/runtime-spec/blob/master/runtime.md)
